Adds support for renderProperties, customNodes, auto-addition of key prop when not present
renderProperties is being used to model props in NavigationExperimental that contain subtrees of components

Custom nodes are ones that cannot be looked up directly in ReactNative - such that when the component tree is being
rendered we need to know which module and name to use to lookup the component that can be passed into React.createElement

key prop auto-addition added when creating elements - prefixed with the name "elm-native-ui-auto-added-". If a key prop
already exists (has been added in the app) it is ignored and the app value used instead. Appears that jsx does something similar
for some components as without it the translated card stack example from UIExplorer displays a warning message