Add initial ListView implementation
Due to the imperative nature of ListView, this implementation doesn't
really work well with the Elm architecture, forcing you to manually keep
a DataSource up to date.

An experimental alternative called
[FlatList](https://medium.com/@cooperka/react-native-new-flatlist-component-30db558c7a5b#.3kodtbtz7)
was recently added to ReactNative which is likely a better long term
fit, but this works in the meantime.

Example use:

    type alias Model =
      { wombats : List Wombat
      , wombatDataSource : DataSource Wombat
      }

    view model =
      listView
        model.wombatDataSource
        wombatView
        [ Ui.style
            [ Style.backgroundColor Color.white
            ]
        ]

    update msg model =
      LoadWombats wombats ->
        let
            wombatDataSource = updateDataSource model.wombatDataSource wombats
        in
          ( { model | wombats = wombats, wombatDataSource = wombatDataSource }, Cmd.none )